Calculate input delta using levenshtein diff
The previous system was too naive. Levenshtein diff can account for
difference at any position in the input.

The library is originally a node.js module. It has to be vendored and
modified to be an ES module instead. I couldn't be bothered to set up
something like browserify or webpack.